feat: hook up starter code to db
- Add `starter_code` column to db
  - Modify actions accordingly
- Enforce editor max length
- Refactor languages.ts to be in Room.ts
